requirements.txt
Pillow==9.2.0
tweepy==3.9.0
tzlocal==2.1
APScheduler==3.0.3
selenium==3.141.0
psycopg2==2.9.3
